Fix narrowing conversion warning

+++ src/sprites.cpp
@@ -29,8 +29,8 @@ Sprite::~Sprite() {}
 
 void Sprite::draw(Graphics &graphics, int x, int y) {
   SDL_Rect destinationRectangel = {x, y,
-                                   this->_sourceRect.w * globals::SPRITE_SCALE,
-                                   this->_sourceRect.h * globals::SPRITE_SCALE};
+                                   static_cast<int>(this->_sourceRect.w * globals::SPRITE_SCALE),
+                                   static_cast<int>(this->_sourceRect.h * globals::SPRITE_SCALE)};
   graphics.blitSurface(this->_spriteSheet, &this->_sourceRect,
                        &destinationRectangel);
 }
